read the (somewhat modified) sunau audio file created by the radio
we have to monkeypatch the sunau module!  ugh.
1) the header size is "too large". okay. whatever.
2) add a metadata extraction function for the header.  fields identified so far:
	date/time recorded => datetime.datetime()
	frequency => str()
	alpha tag => str()

there are a lot more fields in the audio pane of the EZ Scan application GRE ships,
so there might be more data available in different logged samples, or they might be
cross referencing into the frequency database.   This is the important data, though.